+++ src/extension.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
-import { CREATE_STATE_PLACE_HOLDER, NAME_ERROR_MESSAGE,CREATE_REDUCER_PLACE_HOLDER, NAME_REG_EXP, DEFAULT_NAME, STATE_EXTENSION, REDUCER_EXTENSION } from './resources/constants';
+import { CREATE_STATE_PLACE_HOLDER, NAME_ERROR_MESSAGE,CREATE_ACTION_PLACE_HOLDER, CREATE_MIDDLEWARE_PLACE_HOLDER, CREATE_REDUCER_PLACE_HOLDER, NAME_REG_EXP, DEFAULT_NAME, STATE_EXTENSION, REDUCER_EXTENSION, MIDDLEWARE_EXTENSION, ACTION_EXTENSION } from './resources/constants';
 import { getFormattedReducerName, getFormattedStateName, getFilePath } from './resources/utils';
 
 
@@ -16,7 +16,7 @@ export function activate(context: vscode.ExtensionContext) {
 		let nameFieldValidator = new RegExp(NAME_REG_EXP);
 		nameField.placeholder = CREATE_STATE_PLACE_HOLDER;
 		nameField.onDidChangeValue((v) => {
-			nameField.validationMessage = nameFieldValidator.test(v)? NAME_ERROR_MESSAGE : '';
+			nameField.validationMessage = nameFieldValidator.test(v) ? NAME_ERROR_MESSAGE : '';
 		});
 		nameField.onDidAccept(() => {
 			if (nameFieldValidator.test(nameField.value)) {
@@ -24,7 +24,7 @@ export function activate(context: vscode.ExtensionContext) {
 			} else {
 				nameField.hide();
 				var name = nameField.value ? nameField.value : DEFAULT_NAME;
-				createFile(focusedFilePath, name, STATE_EXTENSION,getStateGenCode);
+				createFile(focusedFilePath, name, STATE_EXTENSION, getStateGenCode);
 				nameField.validationMessage = '';
 			}
 		});
@@ -41,7 +41,7 @@ export function activate(context: vscode.ExtensionContext) {
 		let nameFieldValidator = new RegExp(NAME_REG_EXP);
 		nameField.placeholder = CREATE_REDUCER_PLACE_HOLDER;
 		nameField.onDidChangeValue((v) => {
-			nameField.validationMessage = nameFieldValidator.test(v)? NAME_ERROR_MESSAGE : '';
+			nameField.validationMessage = nameFieldValidator.test(v) ? NAME_ERROR_MESSAGE : '';
 		});
 		nameField.onDidAccept(() => {
 			if (nameFieldValidator.test(nameField.value)) {
@@ -49,7 +49,7 @@ export function activate(context: vscode.ExtensionContext) {
 			} else {
 				nameField.hide();
 				var name = nameField.value ? nameField.value : DEFAULT_NAME;
-				createFile(focusedFilePath, name, REDUCER_EXTENSION,getReducerGenCode);
+				createFile(focusedFilePath, name, REDUCER_EXTENSION, getReducerGenCode);
 				nameField.validationMessage = '';
 			}
 		});
@@ -57,15 +57,65 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 
 	context.subscriptions.push(createReducer);
+
+	// Create State Command Registering
+	let createMiddleware = vscode.commands.registerCommand('flutter-redux-gen.createMiddleware', (args) => {
+
+		let focusedFilePath = getFilePath(args.path);
+		let nameField = vscode.window.createInputBox();
+		let nameFieldValidator = new RegExp(NAME_REG_EXP);
+		nameField.placeholder = CREATE_MIDDLEWARE_PLACE_HOLDER;
+		nameField.onDidChangeValue((v) => {
+			nameField.validationMessage = nameFieldValidator.test(v) ? NAME_ERROR_MESSAGE : '';
+		});
+		nameField.onDidAccept(() => {
+			if (nameFieldValidator.test(nameField.value)) {
+				nameField.validationMessage = NAME_ERROR_MESSAGE;
+			} else {
+				nameField.hide();
+				var name = nameField.value ? nameField.value : DEFAULT_NAME;
+				createFile(focusedFilePath, name, MIDDLEWARE_EXTENSION, getMiddlewareGenCode);
+				nameField.validationMessage = '';
+			}
+		});
+		nameField.show();
+	});
+
+	context.subscriptions.push(createMiddleware);
+
+	// Create State Command Registering
+	let createAction = vscode.commands.registerCommand('flutter-redux-gen.createAction', (args) => {
+
+		let focusedFilePath = getFilePath(args.path);
+		let nameField = vscode.window.createInputBox();
+		let nameFieldValidator = new RegExp(NAME_REG_EXP);
+		nameField.placeholder = CREATE_ACTION_PLACE_HOLDER;
+		nameField.onDidChangeValue((v) => {
+			nameField.validationMessage = nameFieldValidator.test(v) ? NAME_ERROR_MESSAGE : '';
+		});
+		nameField.onDidAccept(() => {
+			if (nameFieldValidator.test(nameField.value)) {
+				nameField.validationMessage = NAME_ERROR_MESSAGE;
+			} else {
+				nameField.hide();
+				var name = nameField.value ? nameField.value : DEFAULT_NAME;
+				createFile(focusedFilePath, name, ACTION_EXTENSION, getActionGenCode);
+				nameField.validationMessage = '';
+			}
+		});
+		nameField.show();
+	});
+
+	context.subscriptions.push(createAction);
 }
 
-function createFile(fPath: string, name: string, extention: string,getGenCode:Function) {
+function createFile(fPath: string, name: string, extention: string, getGenCode: Function) {
 	const pathWithFileName = fPath + '/' + name.toLocaleLowerCase() + extention;
 	fs.writeFile(pathWithFileName, getGenCode(name), err => {
 		console.log(err);
-		if(err){
+		if (err) {
 			vscode.window.showInformationMessage('Please check your path. Otherwise file a issue in Git Repo. Let me help.');
-		}else{
+		} else {
 			vscode.window.showInformationMessage(`${name}${extention} Created Successfully.`);
 		}
 	});
@@ -103,13 +153,68 @@ class ${sName}State {
 	  `;
 }
 
-function getReducerGenCode(name:string){
+function getReducerGenCode(name: string) {
 	const sName = getFormattedReducerName(name);
 	return `
 import 'package:redux/redux.dart';
 
 final ${sName}Reducer = combineReducers<AppState>([
 ]);
+	`;
+}
+
+function getMiddlewareGenCode(name: string) {
+	const sName = getFormattedStateName(name);
+	return `
+import 'package:redux/redux.dart';
+
+Middleware<AppState> get${sName}(Repository _repo) {
+	return (Store<AppState> store, action, NextDispatcher dispatch) async {
+	dispatch(action);
+	try {
+		// TODO: Write here your middleware logic and api calls
+	} catch (error) {
+		// TODO: API Error handling
+		print(error);
+	}
+	};
+}
+	`;
+}
+
+function getActionGenCode(name: string) {
+	const sName = getFormattedStateName(name);
+	return `
+import 'package:flutter/material.dart';
+
+class ${sName}Action {
+
+	@override
+	String toString() {
+	return '${sName}Action { }';
+	}
+}
+
+class ${sName}SuccessAction {
+	final int isSuccess;
+
+	${sName}SuccessAction({@required this.isSuccess});
+	@override
+	String toString() {
+	return '${sName}SuccessAction { isSuccess: $isSuccess }';
+	}
+}
+
+class ${sName}FailedAction {
+	final String error;
+
+	${sName}FailedAction({@required this.error});
+
+	@override
+	String toString() {
+	return '${sName}FailedAction { error: $error }';
+	}
+}
 	`;
 }