+++ src/resources/constants.ts
@@ -6,5 +6,7 @@ export const ACTION_EXTENSION = '.action.dart';
 export const DEFAULT_NAME = 'default';
 export const CREATE_STATE_PLACE_HOLDER = 'Enter State Name';
 export const CREATE_REDUCER_PLACE_HOLDER = 'Enter Reducer Name';
+export const CREATE_MIDDLEWARE_PLACE_HOLDER = 'Enter Middleware Name';
+export const CREATE_ACTION_PLACE_HOLDER = 'Enter Action File Name';
 export const NAME_ERROR_MESSAGE = 'Please use name without space. Consider using "_"';
 export const NAME_REG_EXP = /[`0-9 ~!@#$%^&*()|+\-=?;:'",.<>\{\}\[\]\\\/]/;